refactor: fix typos and remove console log
